Add README.md and Fantasy Game rules for 2023
The rules and scoring system for the fantasy game are based on
[The Official Formula 1 Fantasy Game]
(https://fantasy.formula1.com/en/game-rules).

These are used to calculate scores and optimize team selection and
in-season actions. While they are meant to replicate the official
rules and therefore assist with playing the official game, this crate
is in no way associated with Formula 1® and there are no assurances
of rule parity between the rule sets herein and those of the official
fantasy game.