Implement ergast::get::{Resource, Filters}
These types encapsulate the details of Ergast's RESTful API into a
resource that can be requested, and a set of optional filters that
can be applied to the response. The Ergast URL formatting is now
hidden behind these types, which also validate the requests, e.g.
not allowing a `round` filter to be used without a `year` filter.

Add response::{StatusTable, Status} types.

TODO:
  * Add documentation examples for Resource, once the `get_*` API
    has been settled.
  * Add more unit tests for Resource::to_url
  * Add unit tests for validation of Filters, e.g. the `round`
    without `year` case.
  * Add support for Page { limit, offset }