Add `fn_call_width = 100` to rustfmt.toml
With `max_width = 120` and default `use_small_heuristics` the
`fn_call_width` comes out to 72, which is a bit too restrictive,
so we explicitly set it to 100 and `cargo fmt` accordingly.

There are also some comment formatting features that I'd like to use,
but are only a vailable in nightly channels. They are included in
rustfmt.toml, but commented out.