Implement and use ergast::response::LapTime
This enum encapsualtes a possible lap time, or whether no time was set.
This is useful, for example, for qualifying results, where None
indicates that a driver did not take place in a session, but
Some(NoTimeSet) indicates that the driver took part in the session, but
did not set a lap time.

Implement serde deserialization, via DisplayFromStr, and use LapTime in
the q1-q3 fields of QualifyingResult.

TODO:
  * Use LapTime in fastest_lap field
  * Implement proper Err, instead of using panics in LapTime::from_str