Refactor Table enum variants and *Table structs
Rename Table enum variants, removing the *Table suffix. With that,
for example, Table::SeasonTable becomes Table::Seasons. The original
intention was to have a 'pub use Table::*', so as to be able to use
SeasonsTable(...), but Table::Seasons is close enough, doesn't require
the use of 'pub use <variants>', and it better aligns with:
https://rust-lang.github.io/rust-clippy/master/index.html#enum_variant_names

Remove external *Table structs and instead use Table variants with
named fields. Ergast may sometimes return more fields than just the
payload, e.g. fields containing the filters used for a query. Those
aren't currently supported, and are unlikely to be useful and/or
supported in the future. However, it is worth noting that this design
decision makes that potential transition more difficult and, as such,
less likely to be implemented.

These changes are somewhat of a deviation from the names returned in the
JSON response, which is something to be avoided, and they constraint
possible future support expansions. However, the cost is deemed worth
the
benefits in readability, ergonomics, compliance with lint rules, and
compatibility with upcoming use of 'enum-as-inner' to further improve
usability and ergonomics.