Use enum-as-inner to improve ergonomics of Table
Table is an enum in order to express that one and only one table may
be contained in a given response, and to encapsulate all the possible
tables. However, the variants are expected to be used disjointly and
rarely handled in the same method. As such, a common pattern should be
to extract the expected variant or fail out. This can be accomplished
somewhat succinctly with let-else statements, but that is still rather
verbose and repetitive. An even more succinct solution is to provide
getter methods for each variant, which we do here with the help of
the enum-as-inner crate (https://crates.io/crates/enum-as-inner).

Change unit test use cases like the one described above to use the new
getter methods.