Refactor and document main ergast::get methods
Rename the foundational `get_*` methods in ergast::get to `get_response`
and `get_response_page`, in order to convey that they return a full
Response. These are expected to be rarely used by the user, who
should instead use the convenience methods that will be built upon
these. Also change the signature to return a Result<Response, Error>,
forcing the use of ergast::response::Response and providing error
handling with the newly implemented ergast::get::Error type.

Document the new Error type and the `get_*` functions.