Refactor and expand ergast::get pagination tests
Refactor pagination tests to explicitly and more thoroughly test the
`get_response*` functions and the possible MultiPage errors. Also change
them to use Resource::SeasonList which is easier to work with in this
case. These changes are in anticipation of the implementation of new
convenience `get_*` functions, and of switching the existing tests to
use said functions instead of `get_response*`. Once that change is
completed, these tests will stand as the only ones explicitly exercising
the `get_response*` interface.