Simplify ergast::get unit tests with new helpers
Implement the following new helper utilities:

  assert_each_expected_in_actual
  assert_each_get_eq_expected
  assert_not_found

Use these to simplify the implementation of the tests that use the new
convenience APIs. This makes the tests somewhat less explicit and more
terse, but it reduces a lot of code duplication, makes the tests more
mantainable, and arguably makes them clearer. As such, the change is
deemed a net benefit.