Implement new and improved GitHub CI Workflow
The new CI workflow includes more steps for setting up the Rust
toolchain dependencies, so that it can be run in a variety of docker
images, e.g. locally with `act` (https://github.com/nektos/act) and the
default image. It also contains more options, e.g. `-D warnings` to fail
if warnings are reported for build, docs, or clippy. Additionally, it
also now runs rustfmt checks.

Ignored tests were previously not run on CI checks, as they depend on
network and Ergast API availability, which could cause problematic
spurious failures if run on every CI check. Now they are run when the
workflow is triggered on a nightly schedule, which is a better middle
ground between never or always running them.