Move Race.schedule field into variant of Payload
Use new serde feature (https://github.com/serde-rs/serde/pull/2403) to
include untagged Schedule as a variant of Payload, along with the other
tagged variants. This allows the removal of the Race.schedule field,
which would be conditionally populated based on the Resource request,
instead accessing a Schedule via the Race.payload field. This better
represents the behavior that Schedule is returned as one of the possible
Payload variants, depending on the Resource request, and is mutually
exclusive with the other variants.

This also removes the need for the PayloadProxy type, custom
serialization, and the problematic Payload::NoPayload variant.