Make response::Race<> generic on the payload type
Change ergast::response::Race to be a generic struct with the payload
type as a generic parameter, defaulting to the Payload enum. This does
not change anything on the existing code, but allows postprocessing
`get_*` API functions to return `Race`s with concrete payloads when the
expected variant is known, which improves the ergonomics downstream.

Getter functions, with more descriptive names than .payload,  can be
implemented for each instantiation. For example:

  impl Race<QualifyingResult> {
    pub fn qualifying_result(&self) -> &QualifyingResult {
      &self.payload
    }
  }