Move ergast::get::Error, etc. into its own module
Dealing with mismatched error types, even if From<> trait is implemented
for conversions, is not very ergonomic - ? does the .into() conversion,
but sometimes it's necessary to .map_err(|e| e.into()). As such, an
effort will be made to unify f1_data into as few error types as
possible, in order to improve ease of use and ergonomics.

Note:
  One issue is that enum-as-inner does not have an option to return an
  explicit error type, it always just returns the enum itself. It may
  be worth explicitly defining these getters to use proper error types.