Use #[cfg(doc)] for doc-specific use declarations
When a type that is improted via a use declaration is only used for
documentation, then compilation generates an "unused import" warning.
To avoid this warning it is required to pass the explicit full path to
types in the doc comments, which is very verbose. Instead, use
\#[cfg(doc)] for documenation-specific imports, so that they are
available in doc comments and don't generate compilation warnings.