Add `deny` lint groups and allow-by-default lints
Add `deny` lint groups and allowed-by-default lints to f1_data/lib.rs.
Ideally this would be added to the workspace Cargo.toml, but that
features has not been stabilized yet:
https://github.com/rust-lang/cargo/issues/12115

Clippy lints from the `pedantic`, `nursery`, and `restriction` groups
are more finicky and require a combination of opt-in and opt-out
strategies to minimize false positives and verbosity.

Some violations have not yet been fixed, so some lints are temporarily
added as `allow` to not break CI, which runs with `-D warnings`. Devs
can locally change them to `warn` to show the warnings to be fixed.

Also fix some violations generated by the newly added lints.