Change verify_has_one_element_and_extract(Vec<T>)
Change verify_has_one_element_and_extract from taking an Iterator
parameter to taking a Vec<T>. While this is less flexible, it removes
the need to .map(into_iter), making the code less verbose and more
readable. It can always be changed back in the unlikely scenario that
it need to support other containers.