Add `cargo bench` job to GitHub CI Workflow
This job is is not actually intended to benchmark performance, since
the environment that it runs on may not be stable enough to provide
accurate numbers. Rather, it is intended to catch build issues with
`criterion` benchmark code, which is not built by `cargo build` or
`cargo test`. It only runs on scheduled events, currently once nightly,
since it may be too time consuming to run on every push or pull request,
once more benchmarks are added.