Add ergast::resource::Filters::sprint_pos field
This field is similar to Filters::finish_pos, but instead filtering
based on the finishing position in a Sprint event. Change unit tests
accordingly to verify the newly added functionality.

Also add note about the behavior of Filters::sprint_pos and ::finish_pos
with regards to *Result::position. That is, it's only valid when a driver
finishes the event, i.e. *Result::position_text is a numeric value.