Add support for durations with > 99 [seconds]
This is another hotfix related to 7c4a9b1

It seems that Ergast delta durations are always expressed as
[seconds].[subseconds], even if there are more than 59 or 99
seconds. This should only be allowed when specifically parsing
these deltas, but it's being hotfixed here by simply allowing any
Duration to have more than 99 seconds. A proper fix will be
implemented in the upcoming overhaul of time parsing.