Implement benchmarks for ergast::get analysis
Implement benchmarks for the anlysis of ergast::get methods, currently
get_race_results as an example. While benchmarks that include network
requests may not be very precise or consistent, the benchmark of
individual steps and the comparison against the full get_race_results
call is helpful in figuring out where the bottlnecks are. It also
provides a rough metric to identify changes that may or may not have a
performance impact.