Implement custom deserializer for Payload
Implement custom deserializer for Payload in order to improve the
clarity and helpfulness of error messages when there are parsing
failures. Previously, any parsing failures of tagged variants would
result in a Payload::Schedule variant with all fields None, and often
a subsequent cryptic and unhelpful Error::BadVariant. The new
implementation actually results in an Err(...) with useful error
messages.

Add a unit test to verify this functionality.

Make a minor, mostly cosmetic, change to the error messages of
RaceTime's deserialization, regarding 'time' and 'millis' fields.