Introduce RaceResult::NO_NUMBER and custom deser.
RaceResult::NO_NUMBER, set in RaceResult::number field, represents that
no car number was assigned to a race result. This only happend for a few
entries in two races in the 1960s. As such, it's not worth the ergonomic
cost to have the RaceResult::number field be Option<>, and instead this
value will be set for any RaceResult where the entry was not assigned a
car number, which is represented in the Ergast API by an empty string.

Implement `deserialize_possible_no_number` that deserializes an empty
string as RaceResult::NO_NUMBER.

Note:
  This implementation is a a bit concerning, as it could compromise the
  data validation of the f1_data::ergast module, e.g. if new entires
  incorrectly have the .number field as empty strings. However, without
  making changes to the Ergast API itself, I'm not sure anything better
  could be done about it. This problem can be somewhat mitigated by the
  use of `validate_ergast` (with upcoming support for verification of
  known race results with RaceResult::NO_NUMBER).