Flatten response::MrData object into ::Response
The JSON response from the Ergast API contains a single "MRData" object
- the "root" element in previous JSON specifications - that actually
contains all the metadata and data of a response. Previously, this was
represented by the field Response::mr_data and the public `MrData` type.
While this is more accurate to the real JSON response, it does come with
a cost to ergonomics, and is perhaps leaking unnecessary implementation
details. As such, the .mr_data field and MrData type are being flattened
into the top-level `Response` type, with a custom deserialization
implementation handling the de-nesting.