Fix get_qualifying* by adding filters and assets
It appears that Filters::qualifying_pos is not functional in the new
jolpica-f1 API, so we use `season` and/or `driver_id` filters instead.

Add new QUALIFYING_RESULT assets to facilitate the use of different
filters, as mentioned above, along with the required associated assets,
e.g. RACE, RACE_SCHEDULE, etc. Also use these new assets in other
existing tests.

Leave @todo's to add back tests by `qualifying_pos`, if and when that
is fixed in the new jolpica-f1 API.

Add @todo's to add support for "SprintShootout" fields in race
schedules, which seem to be supported in the jolpica-f1 API.

jolpica-f1 is incorrectly reporting some Q1 qualifying result times.
Change these in assets to the incorrect values, to make the tests pass,
and add @todo's to fix them once they are fixed in upstream jolipca-f1