Implement Response::into/as_table_list<T>, etc.
Implement the trait `TableList` to allow for the generic handling of
all `Table` list inner types, associated `Resource` requests, and the
the extraction of the corresponding variants. Implement this trait for
`Season`, `Driver`, `Constructor`, `Circuit`, and `Status`.

Using this trait, implement the following generic methods of `Response`,
to generically handle all the `TableList` types:

    into_table_list<T: TableList>(self) -> Result<Vec<T>>
    into_table_list_single_element<T: TableList>(self) -> Result<T>
    as_table_list<T: TableList>(&self) -> Result<&Vec<T>>
    as_table_list_single_element<T: TableList>(&self) -> Result<&T>

Add convenience aliases to all these functions for all the types that
have implemented the `TableList` trait.

Implement the following generic functions in ergast::get, to generically
handle requests for all the `TableList` types:

    get_table_list<T: TableList>(filters: Filters) -> Result<Vec<T>>
    get_table_list_single_element<T: TableList>(id: T::ID) -> Result<T>

Rewrite all the get_* functions for the `TableList` types as simple
aliases of the above functions.

Make minor naming and documentation changes for SessionResult to match
the new more expansive TableList trait.