Change most "Ergast" references to "jolpica-f1"
The jolpica-f1 API is a drop-in replacement for the now defunct Ergast
API. In its current state, this crate supports either API, despite their
minor differences (see below). However, it will evolve with the
jolpica-f1 API, so change most "Ergast" references to "jolpica-f1" to
reflect the new development direction of the crate.

Ergast and jolpica-f1 API differences:
https://github.com/jolpica/jolpica-f1/blob/main/docs/ergast_differences.md

Some "Ergast" references still remain, either as @todos, or to be
changed in follow-up commits. For example, `jolpica::time::parse_delta`
still references an issue with the Ergast API, which should be
investigated in the new jolpica-f1 API. Also, all references to Ergast
documentation need to be changed to jolpica-f1 in a follow-up commit.