Pull get_response_page impl into new jolpica::get
This new interface is standalone and decoupled from jolpica::agent.
It is also more granular, allowing for a base URL to be passed as well.
This change is made, in part, in preparation for a simplification of
Agent's get_reponse_* interface, and the introduction of options, e.g.
a configurable base URL which would be used in get_response_* calls.

Introduce jolpica::tests::util module to hold common utilities for
jolpica testing, e.g. `retry_http`, `RATE_LIMIT_SLEEP_DURATION`, etc.