Implement Response/Race::as_/to_info() methods
Implement Response/Race::as_/to_info() methods that return a tuple with
a reference to or a copy of, respectively for as_/to_, all the fields
that are info or metadata and not the payload, e.g. Response::table and
Race::payload - also Response::pagination. These allow comparing these
types' info/metadata for equality, which is useful in data validation.

Use ordered_float::OrderedFloat<f64> instead of f64 in response::Circuit
so that it can be used as a key in hash maps. This will be relevant in
upcoming changes that will use Race::as_info, which contains a Circuit,
as keys in a hashmap, indexmap::IndexMap in this case.

Add TODO about improving get_* and _into_* lap timings and pit stops.