Impl workaround for "+-" times in new known_bugs
Introduce a new module crate::jolpica::tests::known_bugs that contains
tests for known bugs and issues, and the associated workarounds. This
partly serves as a collection of @todo items to investigate and
potentially push for fixes in the jolpica-f1 API. If and when any of
these tests start failing, it may indicate that the underlying issue
has been fixed, and the associated workaround can be removed.

Some race times in the jolpica-f1 API seem to be buggy, e.g. 2023, R3,
P13+, non-lapped cars have 'millis' that are lower than P12, and the
'time', expected as a "+hh:mm:ss.sss" string, is instead something like
"+-1:24:07.342" for P15, for example. I don't understand what this is.
For now, implement a workaround where those values, detected as the
'time' field begining with "+-", are parsed as None.