Add "+-" race time workaround to sprint results
SprintResult race times seem to have the same "+-" isse as RaceResult.
See de3dc73 for the original introduction of this workaround.

Some race times in the jolpica-f1 API seem to be buggy, e.g. 2024, R5,
sprint P20, sprint result has 'millis' that is lower than P19, and the
'time', expected as "+hh:mm:ss.sss" string, is instead "+-1:57:34.853".
I don't understand what this is. For now, as a workaround, those values
are being parsed as None for the race time field. If/when this test
fails, we can investigate further.