Implement workaround for the "hh:mm" race time bug
Some lead lap race times seem to be buggy in the jolpica-f1 API. For
example, 1950, R5, P1, the 'time' should be "2:47:26" but is instead
"2:47". It seems that the seconds component is missing, although the
'millis' is correct and contains the seconds. To handle this issue,
we use a regex to detect the "hh:mm" format, verify that it matches
the 'millis' to within 60s, and construct a [`RaceTime::lead`] from
the 'millis' value. This is a temporary workaround, and the issue
should be addressed in upstream jolpica.

Add relevant tests in `jolpica::gests::known_bugs`, and remove one of
the tests which fails now that the workaround is in place.