Implement ridiculous workaround for 2020, R9, P1
2020, R9, P1 race result has incorrect 'millis', off by 1ms, it should
be 8375060. This causes a parsing error as it finds that the 'time' and
'millis' do not match.

To handle this specific known bug, we check for an exact match of the
known incorrect values, and substitutue them with the correct ones.

This is a ridiculous workaround, but it's the only remaining bug that
causes parsing errors for which there isn't a workaround. To avoid
throwing an error for what's likely to be a commonly requested race
result, we do this monstrosity until jolpica-f1 fixes their data. If
and when that happens it should be transparent to users of this crate.