Introduce /scripts and add */test_no_run_docs.sh
Many of the documentation tests are marked with ```no_run since they
make requests to the jolpica API, and so have the same issues as
--ignored tests, but we still want to check that they compile - doc
tests marked with ```ignore are not compiled, unlike regular tests.
Since Rust doesn't have a built-in way to run these tests, as it does
with --ignored tests, the script `scripts/test_no_run_docs.sh` is
provided as a workaround.

This script tests all documentation tests marked with ```no_run. It
temporarily removes the ```no_run markers, applies a patch to add rate
limiting to the jolpica::get::get_response_page function, runs the
documentation tests, and then reverts all changes. Since it may be run
locally with 'act', it first checks that the working directory is clean
before proceeding, in order to avoid losing any uncommitted changes.

Add running this script to GitHub CI workflows.