Enable rate limiting tests to use local jolpica
Modify get* rate limiting tests to use `get_jolpica_test_base_url()` and
so target a local jolpica-f1 server instance if `LOCAL_JOLPICA=1`. This
requires more careful consideration of request latency in the duration
calculation. To facilitate that, introduce `get_request_avg_duration_ms`
which returns an estimated average duration (in ms) of a request to the
jolpica-f1 API, depending on whether it's to a local server instance or
to the real service, ~15ms and ~300ms respectively.