Add workaround for `unused_assignments` after 1.92
After Rust 1.92, there is an `unused_assignments` lint warning being
triggered by the inner fields of [`Table`] variants, e.g.
[`Table::Seasons::seasons`]. I don't understand why it's being
triggered, how to fix it, or why `#[allow(unused_assignments)]` on the
field or enum is not silencing the warning. For now, silencing it at the
smallest scope that works.